+++ src/main/java/ru/iimm/ontology/visualization/tools/cajun/render/tools/ConstantCajun.java
@@ -1,0 +1,14 @@
+package ru.iimm.ontology.visualization.tools.cajun.render.tools;
+
+/**
+ * Константы для пакеджа cajun.
+ * @author Danilov E.Y.
+ *
+ */
+public interface ConstantCajun
+{
+	/**Стиль для target элеентов.*/
+	public static final String NODE_TYPE_TARGET = "tagretStyle";
+	/**Стиль для обычных элеентов.*/
+	public static final String NODE_TYPE_DEFAULT = "defaultStyle";
+}