+++ src/main/java/ru/iimm/ontology/visualization/tools/graphStream/style/StyleGraph.java
@@ -1,0 +1,26 @@
+package ru.iimm.ontology.visualization.tools.graphStream.style;
+
+public class StyleGraph
+{
+	/**Стандартный стиль для графа.*/
+	public static final String GRAPH = "graph{" 
+		    + StyleGraph.GRAPH_STYLE_STARS
+		    + "}";
+	
+	public static final String GRAPH_STYLE_DEFAULT = ""
+			+ "padding: 60px;"
+			+ "fill-color: black;"
+			+"";
+	
+	public static final String GRAPH_STYLE_STARS = ""
+		    + "canvas-color: black;"
+		    + "fill-mode: gradient-vertical;"
+		    + "fill-color: white, white;"
+		    + "padding: 60px;"
+			/*+ "canvas-color: black;"
+		    + "fill-mode: gradient-vertical;"
+		    + "fill-color: black, #004;"
+		    + "padding: 20px;"*/
+			+"";
+
+}