+++ src/main/java/ru/iimm/ontology/visualization/ui/AbstractGUIVisualMethod.java
@@ -1,0 +1,11 @@
+package ru.iimm.ontology.visualization.ui;
+
+/**
+ * Абстрактный класс для GUI визуальных методов.
+ * @author Danilov E.Y.
+ *
+ */
+public abstract class AbstractGUIVisualMethod implements GUIVisualMethodInt
+{	
+
+}