+++ src/main/java/ru/iimm/ontology/visualization/ui/GUIVisualMethodInt.java
@@ -1,0 +1,14 @@
+package ru.iimm.ontology.visualization.ui;
+
+import java.awt.Component;
+
+/**
+ * Интерфейс для GUI визуального метода.
+ * @author Danilov E.Y.
+ *
+ */
+public interface GUIVisualMethodInt extends GUIElementInt
+{
+	/**Возвращает компонент с визуализацией и GUI элементами.*/
+	public Component getGUIComponent();
+}