+++ CHANGELOG.md
@@ -8,9 +8,10 @@
 ## 1.0.1 (2015-07-15)
 ### API Changes
 - Remove classes that use stale library
+- Rename package *visplugin* -> *.visualization.* and *.GUI.* -> *.ui.*
 ### Other
 - Edit Maven set libraries
 - Edit Maven Build
-- Add plugin.xml, viewconfig-mainTab.xml, META-INF for build Protege plugin
-- Add build.xml for build in Ant
-- Add .gitignore
+- Add *plugin.xml*, *viewconfig-mainTab.xml*, *META-INF* for build Protege plugin
+- Add *build.xml* for build in Ant
+- Add *.gitignore*